pyproject.toml
[project]
name = "code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.20.0",
    "asyncpg>=0.30.0",
    "boto3>=1.35.90",
    "fastapi[standard]>=0.115.6",
    "greenlet>=3.1.1",
    "httpx>=0.28.1",
    "mangum>=0.19.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.0",
    "pytest-asyncio>=0.25.0",
    "python-jose>=3.3.0",
    "requests>=2.32.3",
    "ruff>=0.8.4",
    "sqlmodel>=0.0.22",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = ["--import-mode=importlib", ]
